Вынесли данные для подключения к базе
