+++ config.php
@@ -1,0 +1,9 @@
+<?php
+
+$hostname = "localhost";
+
+$username = "root";
+
+$password = "1234";
+
+$database = "todo";
