+++ schema.sql
@@ -14,21 +14,21 @@ CREATE TABLE IF NOT EXISTS projects (
 CREATE TABLE IF NOT EXISTS task (
                                   id INT AUTO_INCREMENT PRIMARY KEY,
                                   user_id INT NOT NULL,
-                                  proeject_id INT NOT NULL,
-                                  dt_add TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+                                  project_id INT NOT NULL,
+                                  date_add TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                                   status BOOL DEFAULT 0,
                                   title VARCHAR(255) NOT NULL,
                                   file_path VARCHAR(255),
-                                  data_exp DATE NULL,
+                                  date_expired DATE NULL,
                                   INDEX(user_id),
                                   INDEX(title),
-                                  INDEX(data_exp)
+                                  INDEX(date_expired)
 
   );
 
 CREATE TABLE IF NOT EXISTS users (
                                    id INT AUTO_INCREMENT PRIMARY KEY,
-                                   dt_reg TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+                                   registration_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                                    email VARCHAR(128) NOT NULL UNIQUE,
                                    name VARCHAR(255) NOT NULL,
                                    password VARCHAR(255) NOT NULL,
@@ -42,5 +42,5 @@ ALTER TABLE projects
   ADD FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE;
 ALTER TABLE task
   ADD FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE,
-	ADD FOREIGN KEY (proeject_id) REFERENCES projects (id) ON DELETE CASCADE;
+	ADD FOREIGN KEY (project_id) REFERENCES projects (id) ON DELETE CASCADE;
